requirements.txt
python-telegram-bot[job-queue]>=20.0
telethon
yt-dlp
python-dotenv
